A better fix for issue #45
This reverts the previous fix and changes the impl of Drop for SpatialRef
to use OSGRelease instead of OSRDestroySpatialReference. This will just
decrement the reference count for the SpatialRef and destroy it if the
count reaches 0.

This resolves a valgrind error that was still being reported after the previous
fix. This fix also has the advantage that it does not change the API for
existing users.