Feature::field returns Result<Option<FieldValue>>
Currently, Feature::field returns Result<FieldValue>, but fields can be
null. When this is the case, the default value for the FieldValue
variant is returned. This makes it difficult to distinguish e.g. a
legitimate Integer 0 value from the absence of a value.

Instead, return a Result<Option<FieldValue>> so that null fields can be
distinguished from default values.