Add an explicit include to gdal-src Cargo.toml
This commit adds an explicit include to the `Cargo.toml` file of
gdal-src to minimize the amount of code uploaded to crates.io. This
reduces the uncompressed package size from >100MB to ~50MB. This is
still larger than what's allowed by crates.io by default, but I know
that it is possible to get an exception from these default limits by
just asking with a good reason. Source code from an existing established
open source project counts as good reason, so that's just something that
needs to be done before publishing the first version.