Exclude development scripts from published package
During a dependency review we noticed that the gdal crate includes various development scripts. These development scripts shouldn't be there as they might, at some point become problematic. As of now they prevent any downstream user from enabling the `[bans.build.interpreted]` option of cargo deny.

I opted for using an explicit include list instead of an exclude list to prevent these files from being included in the published packages to make sure that everything that's included is an conscious choice.

There was also one script in gdal-src that is reported by cargo-deny and
that could be removed. The other script that's reported unfortunally is
required for the build.